+++ src/System.Threading.Tasks.Parallel/src/System/Threading/Tasks/Parallel.cs
@@ -1060,13 +1060,13 @@ namespace System.Threading.Tasks
             // if cancellation is enabled, we need to register a callback to stop the loop when it gets signaled
             CancellationTokenRegistration ctr = (!parallelOptions.CancellationToken.CanBeCanceled)
                             ? default(CancellationTokenRegistration)
-                            : parallelOptions.CancellationToken.Register((o) =>
+                            : parallelOptions.CancellationToken.UnsafeRegister((o) =>
                             {
                                 // Record our cancellation before stopping processing
                                 oce = new OperationCanceledException(parallelOptions.CancellationToken);
                                 // Cause processing to stop
                                 sharedPStateFlags.Cancel();
-                            }, state: null, useSynchronizationContext: false);
+                            }, state: null);
 
             // ETW event for Parallel For begin
             int forkJoinContextID = 0;
@@ -1322,13 +1322,13 @@ namespace System.Threading.Tasks
             // if cancellation is enabled, we need to register a callback to stop the loop when it gets signaled
             CancellationTokenRegistration ctr = (!parallelOptions.CancellationToken.CanBeCanceled)
                             ? default(CancellationTokenRegistration)
-                            : parallelOptions.CancellationToken.Register((o) =>
+                            : parallelOptions.CancellationToken.UnsafeRegister((o) =>
                             {
                                 // Record our cancellation before stopping processing
                                 oce = new OperationCanceledException(parallelOptions.CancellationToken);
                                 // Cause processing to stop
                                 sharedPStateFlags.Cancel();
-                            }, state: null, useSynchronizationContext: false);
+                            }, state: null);
 
             // ETW event for Parallel For begin
             int forkJoinContextID = 0;
@@ -3121,13 +3121,13 @@ namespace System.Threading.Tasks
             // if cancellation is enabled, we need to register a callback to stop the loop when it gets signaled
             CancellationTokenRegistration ctr = (!parallelOptions.CancellationToken.CanBeCanceled)
                             ? default(CancellationTokenRegistration)
-                            : parallelOptions.CancellationToken.Register((o) =>
+                            : parallelOptions.CancellationToken.UnsafeRegister((o) =>
                             {
                                 // Record our cancellation before stopping processing
                                 oce = new OperationCanceledException(parallelOptions.CancellationToken);
                                 // Cause processing to stop
                                 sharedPStateFlags.Cancel();
-                            }, state: null, useSynchronizationContext: false);
+                            }, state: null);
 
             // Get our dynamic partitioner -- depends on whether source is castable to OrderablePartitioner
             // Also, do some error checking.
